fastapi/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic[email]==2.6.1
python-dotenv==1.0.1
mysqlclient==2.2.4
openai==1.12.0
httpx==0.26.0 